SpectralDensities

Documentation for SpectralDensities.

SpectralDensities.AbstractSD

SpectralDensities.OhmicSD

SpectralDensities.OhmicSD

SpectralDensities.PolySD

SpectralDensities.PolySD

SpectralDensities.correlations

SpectralDensities.reorganisation_energy

SpectralDensities.sd

SpectralDensities.sdoverω

SpectralDensities.AbstractSD — Type

abstract type AbstractSD

AbstractSD represents an abstract bath spectral density.

Any subtype of AbstractSD must at least define either sd or sdoverω for the new type.

SpectralDensities.OhmicSD — Type

struct OhmicSD <: AbstractSD

OhmicSD represents an Ohmic spectral density. It is characterized by an amplitude α representing the strength of the Ohmic coupling. That is

\[J(\omega) = \alpha\omega\]

Fields

α::Float64: The amplitude α, indicating the strength of the Ohmic coupling.

SpectralDensities.OhmicSD — Method

OhmicSD(α)

Construct an Ohmic spectral density with amplitude α.

Arguments

α: The amplitude α, indicating the strength of the Ohmic coupling.

Returns

An instance of the OhmicSD struct representing the Ohmic spectral density.

SpectralDensities.PolySD — Type

struct PolySD <: AbstractSD

PolySD represents a polynomial spectral density. It is characterized by an amplitude α representing the strength of the coupling and the polynomial degree n. That is

\[J(\omega) = \alpha\omega^n\]

Fields

α::Float64: The amplitude α, indicating the strength of the coupling.

n::Int: The polynomial degree.

SpectralDensities.PolySD — Method

PolySD(α, n::Int)

Construct a polynomial spectral density with the given amplitude α and degree n.

Arguments

α: The amplitude α, indicating the strength of the coupling.

n::Int: The polynomial degree.

Returns

An instance of the PolySD struct representing the polynomial spectral density.

SpectralDensities.correlations — Method

correlations(J::AbstractSD, τ, β)

Calculate the correlation function for a spectral density J at a given time delay τ and inverse temperature β.

Arguments

J::AbstractSD: The spectral density.

τ: The time delay at which the correlation function is calculated.

β: The inverse temperature.

Returns

The correlation function for the spectral density J at the given time delay τ and inverse temperature β.

SpectralDensities.reorganisation_energy — Method

reorganisation_energy(J::AbstractSD)

Calculate the reorganization energy of a given spectral density J, i.e.

\[\int_0^\infty \frac{J(\omega)}{\omega} \mathrm{d}\omega \]

Arguments

J::AbstractSD: The spectral density.

Returns

The reorganization energy of the spectral density J.

SpectralDensities.sd — Method

sd(J::T, ω) where T <: AbstractSD

Evaluate the spectral density represented by J at a given frequency ω, i.e. J(ω).

Arguments

J::T: The spectral density.

ω: The frequency at which the spectral density is evaluated.

Returns

The spectral density J at the frequency ω.

SpectralDensities.sdoverω — Method

sdoverω(J::T, ω) where T <: AbstractSD

Evaluate the spectral density represented by J divided by a given frequency ω, i.e. J(ω)/ω.

Arguments

J::T: The spectral density.

ω: The frequency at which the spectral density is evaluated.

Returns

The spectral density J(ω) divided by ω.