SpectralDensities

Documentation for SpectralDensities.

SpectralDensities.AbstractSD

SpectralDensities.correlations

SpectralDensities.reorganisation_energy

SpectralDensities.sd

SpectralDensities.sdoverω

SpectralDensities.AbstractSD — Type

abstract type AbstractSD

AbstractSD represents an abstract bath spectral density.

Any subtype of AbstractSD must at least define either sd or sdoverω for the new type.

SpectralDensities.correlations — Method

correlations(J::AbstractSD, τ, β)

Calculate the correlation function for a spectral density J at a given time delay τ and inverse temperature β.

Arguments

J::AbstractSD: The spectral density.

τ: The time delay at which the correlation function is calculated.

β: The inverse temperature.

Returns

The correlation function for the spectral density J at the given time delay τ and inverse temperature β.

SpectralDensities.reorganisation_energy — Method

reorganisation_energy(J::AbstractSD)

Calculate the reorganization energy of a given spectral density J, i.e.

\[\int_0^\infty \frac{J(\omega)}{\omega} \mathrm{d}\omega \]

Arguments

J::AbstractSD: The spectral density.

Returns

The reorganization energy of the spectral density J.

SpectralDensities.sd — Method

sd(J::T, ω) where T <: AbstractSD

Evaluate the spectral density represented by J at a given frequency ω, i.e. J(ω).

Arguments

J::T: The spectral density.

ω: The frequency at which the spectral density is evaluated.

Returns

The spectral density J at the frequency ω.

SpectralDensities.sdoverω — Method

sdoverω(J::T, ω) where T <: AbstractSD

Evaluate the spectral density represented by J divided by a given frequency ω, i.e. J(ω)/ω.

Arguments

J::T: The spectral density.

ω: The frequency at which the spectral density is evaluated.

Returns

The spectral density J(ω) divided by ω.