SpectralDensities

Documentation for SpectralDensities.

SpectralDensities.AbstractSD

SpectralDensities.DebyeSD

SpectralDensities.DebyeSD

SpectralDensities.ExponentialCutoffSD

SpectralDensities.ExponentialCutoffSD

SpectralDensities.GaussianCutoffSD

SpectralDensities.GaussianCutoffSD

SpectralDensities.HardCutoffSD

SpectralDensities.HardCutoffSD

SpectralDensities.LorentzDrudeSD

SpectralDensities.LorentzianSD

SpectralDensities.LorentzianSD

SpectralDensities.OhmicSD

SpectralDensities.OhmicSD

SpectralDensities.OverdampedSD

SpectralDensities.PolySD

SpectralDensities.PolySD

SpectralDensities.UnderdampedSD

SpectralDensities.correlations

SpectralDensities.reorganisation_energy

SpectralDensities.sd

SpectralDensities.sdoverω

SpectralDensities.AbstractSD — Type

abstract type AbstractSD

AbstractSD represents an abstract bath spectral density.

Any subtype of AbstractSD must at least define either sd or sdoverω for the new type.

SpectralDensities.DebyeSD — Type

struct DebyeSD <: AbstractSD

DebyeSD represents a Debye spectral density. It is characterized by an amplitude α representing the strength of the coupling and the cutoff frequency ωc. That is

\[J(\omega) = \frac{2\alpha}{\pi}\frac{\omega\omega_c^2}{\omega^2 + \omega_c^2}\]

Fields

α::Float64: The amplitude α, indicating the strength of the coupling.

ωc::Float64: The cutoff frequency.

SpectralDensities.DebyeSD — Method

DebyeSD(α::Real, ωc::Real)

Construct a Debye spectral density with the given amplitude α and cutoff frequency ωc.

Arguments

α::Real: The amplitude α, indicating the strength of the coupling.

ωc::Real: The cutoff frequency.

Returns

An instance of the DebyeSD struct representing the Debye spectral density.

SpectralDensities.ExponentialCutoffSD — Type

struct ExponentialCutoffSD{T <: AbstractSD} <: AbstractSD

ExponentialCutoffSD represents a spectral density with an exponential frequency cutoff. It is parameterized by the underlying spectral density J, and the frequency cutoff ωcutoff. That is

\[J_\mathrm{exp}(\omega) = J(\omega)e^{-\omega/\omega_c}\]

Fields

J::T: The underlying spectral density.

ωcutoff::Float64: The frequency cutoff.

SpectralDensities.ExponentialCutoffSD — Method

ExponentialCutoffSD(J::AbstractSD, ωcutoff)

Construct a spectral density with an exponential frequency cutoff using the underlying spectral density J and the given frequency cutoff ωcutoff.

Arguments

J::AbstractSD: The underlying spectral density.

ωcutoff: The frequency cutoff.

Returns

An instance of the ExponentialCutoffSD struct representing the spectral density with an exponential frequency cutoff.

SpectralDensities.GaussianCutoffSD — Type

struct GaussianCutoffSD{T <: AbstractSD} <: AbstractSD

GaussianCutoffSD represents a spectral density with a Gaussian frequency cutoff. It is parameterized by the underlying spectral density J, and the frequency cutoff ωcutoff. That is math J_\mathrm{gauss}(\omega) = J(\omega)e^{-\omega^2/\omega_c^2}``

Fields

J::T: The underlying spectral density.

ωcutoff::Float64: The frequency cutoff.

SpectralDensities.GaussianCutoffSD — Method

GaussianCutoffSD(J::AbstractSD, ωcutoff)

Construct a spectral density with a Gaussian frequency cutoff using the underlying spectral density J and the given frequency cutoff ωcutoff.

Arguments

J::AbstractSD: The underlying spectral density.

ωcutoff: The frequency cutoff.

Returns

An instance of the GaussianCutoffSD struct representing the spectral density with a Gaussian frequency cutoff.

SpectralDensities.HardCutoffSD — Type

struct HardCutoffSD{T <: AbstractSD} <: AbstractSD

HardCutoffSD represents a spectral density with a hard frequency cutoff. It is parameterized by the underlying spectral density J, and the frequency cutoff ωcutoff. That is

\[J_\mathrm{hard}(\omega) = J(\omega)\Theta(\omega_c - \omega)\]

where $\Theta$ is the Heaviside theta function.

Fields

J::T: The underlying spectral density.

ωcutoff::Float64: The frequency cutoff.

SpectralDensities.HardCutoffSD — Method

HardCutoffSD(J::AbstractSD, ωcutoff)

Construct a spectral density with a hard frequency cutoff using the underlying spectral density J and the given frequency cutoff ωcutoff.

Arguments

J::AbstractSD: The underlying spectral density.

ωcutoff::Float64: The frequency cutoff.

Returns

An instance of the HardCutoffSD struct representing the spectral density with a hard frequency cutoff.

SpectralDensities.LorentzDrudeSD — Type

struct LorentzDrudeSD <: AbstractSD

Represents a Lorentz-Drude spectral density. This is just an alias for DebyeSD.

SpectralDensities.LorentzianSD — Type

struct LorentzianSD <: AbstractSD

LorentzianSD represents a Lorentzian spectral density. It is characterized by an amplitude α representing the strength of the coupling, the peak centre frequency ω0, and the width of the Lorentzian peak Γ. That is

\[J(\omega) = \frac{\alpha\Gamma}{\pi}\frac{\omega}{(\omega^2 - \omega_0^2)^2 + \omega^2\Gamma^2}\]

Fields

α::Float64: The amplitude α, indicating the strength of the coupling.

ω0::Float64: The centre frequency of the Lorentzian peak.

Γ::Float64: The width of the Lorentzian peak.

SpectralDensities.LorentzianSD — Method

LorentzianSD(α::Real, ω0::Real, Γ::Real)

Construct a Lorentzian spectral density with the given amplitude α, centre frequency ω0, and width Γ.

Arguments

α::Real: The amplitude α, indicating the strength of the coupling.

ω0::Real: The centre frequency of the Lorentzian peak.

Γ::Real: The width of the Lorentzian peak.

Returns

An instance of the LorentzianSD struct representing the Lorentzian spectral density.

SpectralDensities.OhmicSD — Type

struct OhmicSD <: AbstractSD

OhmicSD represents an Ohmic spectral density. It is characterized by an amplitude α representing the strength of the Ohmic coupling. That is

\[J(\omega) = \alpha\omega\]

Fields

α::Float64: The amplitude α, indicating the strength of the Ohmic coupling.

SpectralDensities.OhmicSD — Method

OhmicSD(α::Real)

Construct an Ohmic spectral density with amplitude α.

Arguments

α::Real: The amplitude α, indicating the strength of the Ohmic coupling.

Returns

An instance of the OhmicSD struct representing the Ohmic spectral density.

SpectralDensities.OverdampedSD — Type

struct OverdampedSD <: AbstractSD

Represents an overdamped spectral density. This is just an alias for DebyeSD.

SpectralDensities.PolySD — Type

struct PolySD <: AbstractSD

PolySD represents a polynomial spectral density. It is characterized by an amplitude α representing the strength of the coupling and the polynomial degree n. That is

\[J(\omega) = \alpha\omega^n\]

Fields

α::Float64: The amplitude α, indicating the strength of the coupling.

n::Int: The polynomial degree.

SpectralDensities.PolySD — Method

PolySD(α::Real, n::Int)

Construct a polynomial spectral density with the given amplitude α and degree n.

Arguments

α::Real: The amplitude α, indicating the strength of the coupling.

n::Int: The polynomial degree.

Returns

An instance of the PolySD struct representing the polynomial spectral density.

SpectralDensities.UnderdampedSD — Type

struct UnderdampedSD <: AbstractSD

Represents an underdamped spectral density. This is just an alias for LorentzianSD.

SpectralDensities.correlations — Method

correlations(J::AbstractSD, τ, β)

Calculate the correlation function for a spectral density J at a given time delay τ and inverse temperature β.

Arguments

J::AbstractSD: The spectral density.

τ: The time delay at which the correlation function is calculated.

β: The inverse temperature.

Returns

The correlation function for the spectral density J at the given time delay τ and inverse temperature β.

SpectralDensities.reorganisation_energy — Method

reorganisation_energy(J::AbstractSD)

Calculate the reorganization energy of a given spectral density J, i.e.

\[\int_0^\infty \frac{J(\omega)}{\omega} \mathrm{d}\omega \]

Arguments

J::AbstractSD: The spectral density.

Returns

The reorganization energy of the spectral density J.

SpectralDensities.sd — Method

sd(J::T, ω) where T <: AbstractSD

Evaluate the spectral density represented by J at a given frequency ω, i.e. J(ω).

Arguments

J::T: The spectral density.

ω: The frequency at which the spectral density is evaluated.

Returns

The spectral density J at the frequency ω.

SpectralDensities.sdoverω — Method

sdoverω(J::T, ω) where T <: AbstractSD

Evaluate the spectral density represented by J divided by a given frequency ω, i.e. J(ω)/ω.

Arguments

J::T: The spectral density.

ω: The frequency at which the spectral density is evaluated.

Returns

The spectral density J(ω) divided by ω.